refactor(host-contracts): devnet canonical deploy executes the DAO payload
Pilot of the unified upgrade-path principle: the direct (private-key) path is
now minimal and built on the DAO path instead of running parallel machinery.

- new tasks/utils/daoUpgrade.ts hosts the generic prepare/print/verify
  machinery (moved from taskMigrate) plus executePreparedDaoUpgrade, the
  entire direct path: send the prepared upgradeToAndCall payload with the
  deployer key
- prepareCanonicalSnapshotUpgrade (snapshot -> PreparedDaoUpgrade) is shared
  by task:prepareDeployProtocolConfigFromCanonical (print for the DAO) and
  task:deployProtocolConfigFromCanonical (execute immediately) — devnet runs
  byte-identical calldata to what the DAO signs, so every devnet deploy
  exercises the DAO payload end to end
- applyCanonicalSnapshot (parallel upgradeProxy machinery) deleted
